move resample and subset transformations to variables
These are operations required to align variable with a common set of dimension to store them in a model inputs datastore. Previously these methods were utilities in the modules repo, but given they require detailed knowledge of the underlying data structure of a variable, it makes more sense to have them here.